Define shims for stack probes using naked functions with RFC 2873 style asm!.
This should fix #1722.
